backend/requirements.txt
Django==4.0.0
djangorestframework==3.13.0
djangorestframework-simplejwt==5.0.0
django-mptt==0.13.4
psycopg2-binary==2.9.3
gunicorn==20.1.0
django-cors-headers==3.10.0
Pillow==9.0.0
python-dotenv==0.19.2
cos-python-sdk-v5
requests==2.27.1

